requirements.txt
asgiref==3.2.7
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.7
django-admin-interface==0.12.2
django-ckeditor==5.9.0
django-colorfield==0.3.1
django-debug-toolbar==2.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.3.0
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.2
psycopg2-binary==2.8.5
pylint==2.4.4
python-decouple==3.3
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
